updated Splitter test for TrafficAllocation
