+++ phpunit.xml.dist
@@ -30,6 +30,9 @@
         <testsuite name="engine">
             <directory suffix="Test.php">tests/Suite/Engine/</directory>
         </testsuite>
+        <testsuite name="traffic_allocation">
+            <directory suffix="Test.php">tests/Suite/TrafficAllocation/</directory>
+        </testsuite>
         <testsuite name="input_validation">
             <directory suffix="Test.php">tests/Suite/InputValidation/</directory>
         </testsuite>
@@ -55,6 +58,7 @@
             <directory suffix="Test.php">tests/Suite/Component/</directory>
             <directory suffix="Test.php">tests/Suite/DynamicConfigurations/</directory>
             <directory suffix="Test.php">tests/Suite/Engine/</directory>
+            <directory suffix="Test.php">tests/Suite/TrafficAllocation/</directory>
             <directory suffix="Test.php">tests/Suite/InputValidation/</directory>
             <directory suffix="Test.php">tests/Suite/Matchers/</directory>
             <directory suffix="Test.php">tests/Suite/Metrics/</directory>
